Fixed test path and files charset
