+++ .gitignore
@@ -1,3 +1,4 @@
+.vscode
 # Ignore Gradle project-specific cache directory
 .gradle
