--- app/src/main/java/serverest/userFlowTest/Exercicio04Test.java
@@ -1,2 +1,0 @@
-package serverest.userFlowTest;public class Exercicio04Test {
-}
