--- app/src/main/java/serverest/util/Mensagem.java
@@ -1,9 +1,0 @@
-package serverest.util;
-
-public class Mensagem {
-
-    public static final String cadastroSucesso = "Cadastro realizado com sucesso";
-    public static final String emailJaUsado = "Este email já está sendo usado";
-    public static final String loginSucesso = "Login realizado com sucesso";
-    public static final String deleteProdutoSucesso = "Registro excluído com sucesso. Estoque dos produtos reabastecido";
-    public static final String deleteUsuarioSucesso = "Usuário deletado com sucesso"
