+++ app/src/test/java/serverest/userFlowTest/Exercicio01Test.java
@@ -1,0 +1,13 @@
+package serverest.userFlowTest;
+
+import org.junit.Test;
+import serverest.control.Usuario;
+
+public class Exercicio01Test {
+
+    @Test
+    public void getUsuarios() {
+        Usuario.listarUsuarios();
+    }
+}
+
