+++ app/src/test/java/serverest/userFlowTest/Exercicio04Test.java
@@ -9,27 +9,22 @@ import serverest.model.ProdutoDTO;
 import serverest.model.UsuarioDTO;
 import serverest.util.Mensagem;
 
-public class Exercicio03Test {
-
+public class Exercicio04Test {
     @Test
-    public void validarEstoque() {
+    public void validarUsuario()
+
+    {
         UsuarioDTO usuarioDTO = new UsuarioDTO("true");
-        //Cadastrar um usu·rio
+        //Cadastrar um usu√°rio
         String userId = Usuario.cadastrarUsuario(usuarioDTO, HttpStatus.SC_CREATED, Mensagem.cadastroSucesso);
-        //Autenticar usu·rio
+        //Autenticar usu√°rio
         String userToken = Usuario.autenticarUsuario(usuarioDTO, HttpStatus.SC_OK, Mensagem.loginSucesso);
         //Cadstrar um produto
         ProdutoDTO produtoDTO = Produto.gerarPokemon(100, 100);
         String productId = Produto.cadastrarProduto(produtoDTO, userToken, HttpStatus.SC_CREATED, Mensagem.cadastroSucesso);
-        //Verificar estoque do produto
-        Produto.verificarEstoque(productId, 100);
         //Cadastrar carrinho
         Carrinho.cadastrarCarrinho(productId, 1, userToken, HttpStatus.SC_CREATED, Mensagem.cadastroSucesso);
-        //Verificar estoque do produto
-        Produto.verificarEstoque(productId, 99);
-        //Cancelar a compra
-        Carrinho.cancelarCompra(userToken, HttpStatus.SC_OK, Mensagem.deleteProdutoSucesso);
-        //Verificar estoque do produto
-        Produto.verificarEstoque(productId, 100);
+        //Verificar deletar usuario
+
     }
 }
